+++ CHANGELOG.md
@@ -7,6 +7,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased - 2023-09-08
+
+## Added
+
+- Can use the `addon` property for the `Autocomplete` component
+
+## Fixed
+
+- The `Autocomplete` component now takes on the full width of its parent element
+
 ## 7.0.2 - 2023-08-30
 
 =======